+++ front/src/component/MarkdownViewer/MarkdownViewer.tsx
@@ -1,26 +1,10 @@
 import React from 'react';
-import cb from 'react-syntax-highlighter/dist/esm/styles/prism/cb';
-
-import { SyntaxWrapper, MarkdownWrapper, InlineCode } from './styled';
+import { MarkdownWrapper } from './styled';
+import renderers from './renderers';
 
 interface Props {
   content: string;
 }
-
-interface RenderersProps {
-  language?: string;
-  value: string;
-}
-
-const renderers = {
-  code: ({ language, value }: RenderersProps) => (
-    <SyntaxWrapper style={cb} language={language}>
-      {value}
-    </SyntaxWrapper>
-  ),
-
-  inlineCode: ({ value }: RenderersProps) => <InlineCode>{value}</InlineCode>,
-};
 
 const MarkdownViewer = ({ content }: Props) => {
   return <MarkdownWrapper renderers={renderers}>{content}</MarkdownWrapper>;
