+++ front/src/component/MarkdownViewer/renderers/index.ts
@@ -1,0 +1,13 @@
+import code from './code';
+import inlineCode from './inlineCode';
+import blockquote from './blockquote';
+import heading from './heading';
+
+const renderers = {
+  code,
+  inlineCode,
+  blockquote,
+  heading,
+};
+
+export default renderers;
