+++ front/src/component/MarkdownViewer/renderers/inlineCode.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import styled from 'styled-components';
+
+interface InlineCodeProps {
+  value: string;
+}
+
+const inlineCode = ({ value }: InlineCodeProps) => <InlineCode>{value}</InlineCode>;
+
+const InlineCode = styled.code`
+  background-color: ${({ theme }) => theme.CODE_BACKGROUND};
+  color: ${({ theme }) => theme.CODE_INLINE};
+  padding: 2px 3px;
+`;
+
+export default inlineCode;
